third_party/WebKit/LayoutTests/animations/custom-properties/registered-var-to-registered-any-order.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<div id="target"></div>
<script>
setup(() => {
  for (let name of ['--a', '--b']) {
    CSS.registerProperty({
      name,
      syntax: '<number>',
      initialValue: '0',
    });
  }
});

test(() => {
  // This test runs two animations, each with a different dependency ordering
  // between --a and --b, to check that the lexicographical order of the custom
  // property names don't influence the dependency resolution ordering.
  let a = target.animate({'--a': ['100', 'var(--b)']}, 100);
  let b = target.animate({'--b': ['200', '200']}, 100);

  a.currentTime = 50;
  assert_equals(getComputedStyle(target).getPropertyValue('--a'), '150', '--a animating towards --b');
  assert_equals(getComputedStyle(target).getPropertyValue('--b'), '200', '--b stationary at 200');

  a.cancel();
  b.cancel();

  a = target.animate({'--a': ['200', '200']}, 100);
  b = target.animate({'--b': ['100', 'var(--a)']}, 100);

  b.currentTime = 50;
  assert_equals(getComputedStyle(target).getPropertyValue('--a'), '200', '--a stationary at 200');
  assert_equals(getComputedStyle(target).getPropertyValue('--b'), '150', '--b animating towards --a');
}, 'Registered custom property animation keyframes with var() references should not be affected by lexicographical property name ordering');
</script>
